requirements.txt
pandas
matplotlib
requests
python-dotenv
